feat(backend): Improve code clarity, maintainability, and error handling
This refactor introduces significant improvements across various backend modules, focusing on code clarity, maintainability, and error handling, particularly around cryptography and API interactions.

Key changes include:
- **Enhanced Readability & Conciseness:** Adoption of modern Rust idioms (`let-else`, `map_or_else`, `Self::`, `format!`, `#\[must_use\]`), type aliases for complex results, and simplified string operations.
- **Robust Error Handling:** Centralized error logging and conversion for `genai::Error`, consistent error message formatting, and explicit validation for required fields.
- **Optimized Performance & Efficiency:** Elimination of redundant `clone()` calls, use of `push` for single characters, and more efficient `Option` handling.
- **Increased Modularity & Testability:** Breakdown of large functions into smaller, focused helpers (e.g., `delete_all_all_sessions_for_user`), and improved organization of tests.
- **Improved Concurrency & Safety:** Explicit scoping of `RwLockWriteGuard` to ensure prompt lock release.
- **Streamlined Configuration:** Transition of critical configuration fields from `Option<String>` to `String` where values are always expected, simplifying access.
- **Enhanced Code Hygiene:** Extensive removal of unnecessary `#\[allow(dead_code)\]` attributes and unused imports, indicating a cleaner and more actively used codebase.

These changes collectively contribute to a more robust, efficient, and easier-to-understand backend.

+++ backend/src/services/history_manager.rs
@@ -424,7 +424,7 @@ mod tests {
 
         // Limit allows only part of the first message, second message fully
         let limit = 20;
-        let managed = manage_history(history, "truncate_tokens", limit as i32);
+        let managed = manage_history(history, "truncate_tokens", limit);
 
         assert_eq!(
             managed.len(),
